+++ app/Http/Controllers/Admin/SosmedController.php
@@ -89,7 +89,7 @@ class SosmedController extends Controller
                 $path = $file->storeAs('sosmed', $imageName, 'public');
                 $validated['image_path'] = $path;
                 // $validated['image_thumb'] = $path;
-                Storage::disk('public')->delete($sosmed->image_path);
+                $sosmed->image_path && Storage::disk('public')->delete($sosmed->image_path);
             }
             $sosmed->update($validated);
             return response()->json(new ResponseSuccess($sosmed, "Success", "Success Update Sosmed"));
@@ -107,7 +107,7 @@ class SosmedController extends Controller
     {
         try {
             $sosmed->delete();
-            Storage::disk('public')->delete($sosmed->image_path);
+            $sosmed->image_path && Storage::disk('public')->delete($sosmed->image_path);
             return response()->json(new ResponseSuccess($sosmed, "Success", "Success Update Sosmed"));
         } catch (\Throwable $th) {
             Log::error($th->getMessage());
